Make the GNU Info directory node stateless
The info program looks for it in /usr/share/info/dir, which our
statelessness requires to be an immutable file. So make it immutable: a
symlink to /var/lib/info/dir, which is managed by a systemd service that
is started whenever /usr/share/info is modified (no swupd trigger
required). This service is enabled by default.

This requires clearlinux/autospec#540 to install the script.

Fixes clearlinux/distribution#1111.